Refactor createRequestBody to use default options
The createRequestBody function has been refactored to use defaultTranslateOptions for its parameters. Where previously passed as an argument, "rpcids", "from" and "to" properties now default to values from defaultTranslateOptions unless otherwise specified. This leads to cleaner code and eliminates the need for explicit selection every time the function is called.